Parry Reader update
moved functions to separate code
