+++ Project Ronin/Assets/Scripts/Ability System/ParryReader.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 308a0eb4b515d904a880f3f0db7b30cc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
